anemia-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
numpy==2.1.3
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
pillow==11.0.0
python-multipart==0.0.20
huggingface-hub==0.26.5
